+++ server/src/main/java/com/github/AllenDuke/producerService/RPCServerHandler.java
@@ -64,8 +64,7 @@ public class RPCServerHandler extends SimpleChannelInboundHandler {
             }
             log.error("该远程客户端一天内调用错误次数达到100次，认为已受到了该客户端的攻击，" +
                     "24小时内不允许再次连接，将告知客户端并断开连接");
-            ServerMessage serverMessage=new ServerMessage(0,0,false,
-                    "请24小时之后再试！");
+            ServerMessage serverMessage=new ServerMessage((short) 0,0,false, "请24小时之后再试！");
             ctx.writeAndFlush(JSON.toJSONString(serverMessage));
             ctx.close();
             return;
@@ -92,8 +91,7 @@ public class RPCServerHandler extends SimpleChannelInboundHandler {
             clientMessage = JSON.parseObject((String) msg, ClientMessage.class);
         }catch (Exception e){
             log.error("解析异常，放弃本次解析任务，即将通知客户端",e);
-            ServerMessage serverMessage=new ServerMessage(0
-                    ,0, false,"服务器解析异常");
+            ServerMessage serverMessage=new ServerMessage((short) 0,0, false,"服务器解析异常");
             ctx.writeAndFlush(JSON.toJSONString(serverMessage));
             recordInvokeException(ctx,e);
             return;
